+++ static_precompiler/compilers/babel.py
@@ -1,6 +1,7 @@
 import json
 import os
 import posixpath
+import warnings
 
 from static_precompiler import exceptions, utils
 
@@ -17,22 +18,33 @@ class Babel(base.BaseCompiler):
     input_extension = "es6"
     output_extension = "js"
 
-    def __init__(self, executable="babel", sourcemap_enabled=False, modules=None):
+    def __init__(self, executable="babel", sourcemap_enabled=False, modules=None, plugins=None):
         self.executable = executable
         self.is_sourcemap_enabled = sourcemap_enabled
+        if modules:
+            warnings.warn("'modules' option is removed in Babel 6.0. Use `plugins` instead.", DeprecationWarning)
         self.modules = modules
+        self.plugins = plugins
         super(Babel, self).__init__()
+
+    def get_extra_args(self):
+        args = []
+
+        if self.modules is not None:
+            args += ["--modules", self.modules]
+
+        if self.plugins is not None:
+            args += ["--plugins", self.plugins]
+
+        return args
 
     def compile_file(self, source_path):
         args = [
             self.executable,
-        ]
+        ] + self.get_extra_args()
 
         if self.is_sourcemap_enabled:
             args.append("-s")
-
-        if self.modules is not None:
-            args.extend(["--modules", self.modules])
 
         full_output_path = self.get_full_output_path(source_path)
 
@@ -66,11 +78,8 @@ class Babel(base.BaseCompiler):
 
     def compile_source(self, source):
         args = [
-            self.executable
-        ]
-
-        if self.modules is not None:
-            args.extend(["--modules", self.modules])
+            self.executable,
+        ] + self.get_extra_args()
 
         out, errors = utils.run_command(args, source)
         if errors: